Added persistence (and update) of 'averageHashDuration' in NodeState (via GenericFileCachedUrlSet. Changed default hashtest length from 1 second to 10 seconds in SystemMetrics. Named the TreeWalkThread.

+++ test/src/org/lockss/daemon/TestSystemMetrics.java
@@ -74,7 +74,7 @@ public class TestSystemMetrics extends LockssTestCase {
     assertEquals(estimate, estimate2);
   }
 
-  public static void configHashParams(int duration, int step)
+  public static void configHashParams(long duration, int step)
       throws IOException {
     String s = SystemMetrics.PARAM_HASH_TEST_DURATION + "=" + duration;
     String s2 = SystemMetrics.PARAM_HASH_TEST_BYTE_STEP + "=" + step;
